requirements.txt
Jinja2
requests
python-benedict
markdown